lib: decode `die_id` from integer value
This patch includes an optimization when the die id name is parsed.
Instead of converting an `u8` value to `String` and then getting the
name of the die from the JSON, this patch includes a `Map<u8, String>`

The header library can get the die name from this new `Map<u8, String>`
and removes the need to convert the original `u8` to `String`.

Signed-off-by: Rivas Paz, Jose L <jose.l.rivas.paz@intel.com>