lib: include die name decode in `basic_decode`
This patch enables an alternative flow inside `decode` function.

When the current Crash Log record contains a Header Type 6 (which is the
only header at the moment with `die_id` support) and there are no
defined decode definitions, the current `basic_decode` function will
still try to rename the die name when the header is decoded.

This patch also changes the return value of the `decode` function in the
Record from a Result to a Node.

Signed-off-by: Rivas Paz, Jose L <jose.l.rivas.paz@intel.com>