lib: refactor logic for adding `custom_root` to decoded registers
This patch refactors the logic that adds a custom root path to the
decoded Crash Log record.
With this change it doesn't matter which type of decoder was executed,
it could be the generic record or the core decoder, the specific root
path will by applied if the record contains the needed data for the
custom root path or if the data is part of the context of the record.

Signed-off-by: Rivas Paz, Jose L <jose.l.rivas.paz@intel.com>