lib: handle empty regions
The box record payload does not always contain data. Checking if the
region is empty during its construction prevents the invalid box payload
to appear in the region list stored in the Crash Log struct.

This also prevents the unpack command to report empty regions and the
decode function to consider empty files as valid.

    $ touch empty.crashlog
    $ iclg decode empty.crashlog
    Fatal Error: The Crash Log Region is not populated

Signed-off-by: Surply, Pierre <pierre.surply@intel.com>